refactor: 이벤트 코드 내부 패키지명 변경
이전되기 전의 패키지명을 갖고 있어서 의존성 에러 발생함.

#17
