+++ src/main/java/com/study/ticketservice/domain/event/controller/response/EventStatusResponse.java
@@ -1,6 +1,7 @@
-package com.study.ticketservice.event.application.dto.response;
+package com.study.ticketservice.domain.event.controller.response;
 
-import com.study.ticketservice.event.domain.EventStatus;
+import com.study.ticketservice.domain.event.enums.EventStatus;
+
 import java.time.LocalDateTime;
 
 /**
