+++ src/main/java/com/study/ticketservice/domain/event/entity/SeatEntity.java
@@ -1,7 +1,6 @@
-package com.study.ticketservice.event.infrastructure;
+package com.study.ticketservice.domain.event.entity;
 
-import com.study.ticketservice.event.domain.Seat;
-import com.study.ticketservice.event.domain.SeatStatus;
+import com.study.ticketservice.domain.event.enums.SeatStatus;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
