+++ src/main/java/com/study/ticketservice/domain/event/enums/EventStatus.java
@@ -1,4 +1,4 @@
-package com.study.ticketservice.event.domain;
+package com.study.ticketservice.domain.event.enums;
 
 /**
  * 이벤트 상태를 나타내는 열거형
