+++ src/main/java/com/study/ticketservice/domain/event/service/SeatWriterImpl.java
@@ -1,7 +1,8 @@
-package com.study.ticketservice.event.infrastructure;
+package com.study.ticketservice.domain.event.service;
 
-import com.study.ticketservice.event.application.SeatWriter;
-import com.study.ticketservice.event.domain.Seat;
+import com.study.ticketservice.domain.event.entity.Seat;
+import com.study.ticketservice.domain.event.entity.SeatEntity;
+import com.study.ticketservice.domain.event.repository.SeatJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
